% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.logspline-package}
\alias{mlr3learners.logspline}
\alias{mlr3learners.logspline-package}
\title{mlr3learners.logspline: Learners from the {logspline} package for 'mlr3'}
\description{
Adds learners from the {logspline} package to
    'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
